Clean up auth test utils
