+++ libs/test-utils/src/auth.ts
@@ -1,0 +1,48 @@
+import {
+  CardlessVoterUser,
+  ElectionManagerUser,
+  PollWorkerUser,
+  SystemAdministratorUser,
+} from '@votingworks/types';
+
+export function fakeSystemAdministratorUser(
+  props: Partial<SystemAdministratorUser> = {}
+): SystemAdministratorUser {
+  return {
+    role: 'system_administrator',
+    passcode: '123456',
+    ...props,
+  };
+}
+
+export function fakeElectionManagerUser(
+  props: Partial<ElectionManagerUser> = {}
+): ElectionManagerUser {
+  return {
+    role: 'election_manager',
+    electionHash: 'election-hash',
+    passcode: '123456',
+    ...props,
+  };
+}
+
+export function fakePollWorkerUser(
+  props: Partial<PollWorkerUser> = {}
+): PollWorkerUser {
+  return {
+    role: 'poll_worker',
+    electionHash: 'election-hash',
+    ...props,
+  };
+}
+
+export function fakeCardlessVoterUser(
+  props: Partial<CardlessVoterUser> = {}
+): CardlessVoterUser {
+  return {
+    role: 'cardless_voter',
+    ballotStyleId: 'fake-ballot-style-id',
+    precinctId: 'fake-precinct-id',
+    ...props,
+  };
+}
