+++ libs/test-utils/src/index.ts
@@ -1,5 +1,6 @@
 export * from './advance_timers';
 export * from './arbitraries';
+export * from './auth';
 export * from './cast_vote_record';
 export * from './child_process';
 export * from './compressed_tallies';
@@ -11,8 +12,6 @@ export * from './has_text_across_elements';
 export * from './mock_function';
 export * from './mock_of';
 export * from './observable';
-export * from './smartcards';
-export * from './smartcard_auth';
 export * from './strip_election_hash';
 export * from './types';
 export * from './zip';
