--- libs/test-utils/src/smartcard_auth/index.ts
@@ -1,3 +1,0 @@
-export * from './auth';
-export * as Inserted from './inserted';
-export * as Dipped from './dipped';
